+++ packages/core/src/hooks/index.ts
@@ -1,2 +1,3 @@
+export * from "./cursor-hooks.js";
 export * from "./dispatcher.js";
 export * from "./scope.js";